% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordGB.R
\name{concordGB}
\alias{concordGB}
\title{concordGB}
\usage{
concordGB(sourcevar, origin = "GB2011", destination = "isic4")
}
\arguments{
\item{sourcevar}{Vector which contains the codes to be converted.}

\item{origin}{Coding scheme of origin (name enclosed in quotes "").}

\item{destination}{Coding scheme of destination (name enclosed in quotes "").}
}
\value{
Returns a vector of concorded codes.
}
\description{
Product Concordance
}
\details{
Now \code{GB2011} / \code{GB2017} can be translated to
  \code{isic4}, and vice varsa. The transition among \code{GB2002},
  \code{GB2011} and \code{GB2017} is also available. A product code is 4 digits.

  R package \code{concordance} provides a concordance among HS, HS0, HS1, HS2, HS3, HS4, ISIC2,
   ISIC3, SITC1, SITC2, SITC3, SITC4, BEC, NAICS and SIC.
}
\examples{
# translate GB2011 codes to isic4 codes
concordGB(c('0142','2411'))
}
