% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GB171102.R
\docType{data}
\name{GB171102}
\alias{GB171102}
\title{GB Code for 2002, 2011 and 2017}
\format{
A data frame with 1527 rows and 6 variables:
\describe{
\item{GB2011}{GB code for 2011}
\item{GB2017}{GB code for 2017}
\item{DesGB2017}{GB code descriptions for 2017}
\item{DesGB2011}{GB code descriptions for 2011}
\item{GB2002}{GB code for 2002}
\item{DesGB2002}{GB code descriptions for 2002}
}
}
\usage{
GB171102
}
\description{
A dataset containing the GB code and their description
}
\keyword{datasets}
